git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@5793 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/GRANULAR/atom_vec_granular.h
@@ -58,7 +58,7 @@ class AtomVecGranular : public AtomVec {
   int data_atom_hybrid(int, char **);
   void data_vel(int, char **);
   int data_vel_hybrid(int, char **);
-  double memory_usage();
+  bigint memory_usage();
 
  private:
   double PI;
