+++ src/atom_vec_atomic.h
@@ -48,7 +48,7 @@ class AtomVecAtomic : public AtomVec {
   int unpack_restart(double *);
   void create_atom(int, double *);
   void data_atom(double *, int, char **);
-  double memory_usage();
+  bigint memory_usage();
 
  private:
   int *tag,*type,*mask,*image;
