+++ src/atom_vec_hybrid.cpp
@@ -842,9 +842,9 @@ void AtomVecHybrid::data_vel(int m, char **values)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-double AtomVecHybrid::memory_usage()
+bigint AtomVecHybrid::memory_usage()
 {
-  double bytes = 0.0;
+  bigint bytes = 0;
   for (int k = 0; k < nstyles; k++) bytes += styles[k]->memory_usage();
   return bytes;
 }
