+++ src/integrate.h
@@ -14,7 +14,6 @@
 #ifndef LMP_INTEGRATE_H
 #define LMP_INTEGRATE_H
 
-#include "lmptype.h"
 #include "pointers.h"
 
 namespace LAMMPS_NS {
@@ -29,7 +28,7 @@ class Integrate : protected Pointers {
   virtual void run(int) = 0;
   virtual void cleanup() {}
   virtual void reset_dt() {}
-  virtual double memory_usage() {return 0.0;}
+  virtual bigint memory_usage() {return 0;}
 
  protected:
   int eflag,vflag;                  // flags for energy/virial computation
