+++ src/modify.h
@@ -14,7 +14,6 @@
 #ifndef LMP_MODIFY_H
 #define LMP_MODIFY_H
 
-#include "lmptype.h"
 #include "stdio.h"
 #include "pointers.h"
 
@@ -96,7 +95,7 @@ class Modify : protected Pointers {
   int read_restart(FILE *);
   void restart_deallocate();
 
-  double memory_usage();
+  bigint memory_usage();
 
  private:
